chore: Move all remaining actions into links
