+++ app/client/packages/rts/src/ctl/backup/links/DiskSpaceLink.ts
@@ -1,5 +1,14 @@
-import { checkAvailableBackupSpace, getAvailableBackupSpaceInBytes } from "..";
 import type { Link } from ".";
+import * as Constants from "../../constants";
+import fsPromises from "fs/promises";
+
+export function checkAvailableBackupSpace(availSpaceInBytes: number) {
+  if (availSpaceInBytes < Constants.MIN_REQUIRED_DISK_SPACE_IN_BYTES) {
+    throw new Error(
+        "Not enough space available at /appsmith-stacks. Please ensure availability of at least 2GB to backup successfully.",
+    );
+  }
+}
 
 export class DiskSpaceLink implements Link {
   async preBackup() {
@@ -8,4 +17,12 @@ export class DiskSpaceLink implements Link {
 
     checkAvailableBackupSpace(availSpaceInBytes);
   }
+}
+
+export async function getAvailableBackupSpaceInBytes(
+    path: string,
+): Promise<number> {
+  const stat = await fsPromises.statfs(path);
+
+  return stat.bsize * stat.bfree;
 }
